backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
PyMuPDF==1.24.14
httpx==0.28.1
pgvector==0.2.5
aiofiles==24.1.0
